+++ .github/workflows/CHANGELOG.md
@@ -1,6 +1,7 @@
 # k8f 
 ## Changes
 - Improved commands descriptions
+- Added Merge Flag to combine new K8S and already existed K8S in kubeconfig file
 ## Bugfix
 ## Known Issue
 - connect command with "dry-run" flag will default to Json insted of Yaml